feat: move PredicateBooleanReturn into an opt-in Gusto/Sorbet sub-department
Renames Gusto/SorbetPredicateBooleanReturn to Gusto/Sorbet/PredicateBooleanReturn
and gates it behind a new config/sorbet.yml, following the existing rails.yml and
sidekiq.yml pattern: the cop's full entry stays in gusto_cops.yml with
`Enabled: false`, and the opt-in file flips it to `Enabled: true`. `init` adds
config/sorbet.yml to a project's inherit_gem list when it sees a sorbet gem in the
Gemfile or Gemfile.lock, so projects that don't use Sorbet never see the cop.

Encoding "this is a Sorbet rule" in the department rather than the cop name leaves
room for the other Sorbet cops we want to incubate here before proposing them to
rubocop-sorbet, and it stops the name from reading as one flat Gusto rule.

Also fixes a hole this change would otherwise have opened in the config validation
specs: they collected cops via `Registry#with_department(:Gusto)`, whose department
for a nested cop is `:"Gusto/Sorbet"`, not `:Gusto`. Every Gusto/* cop in a
sub-department would have silently skipped the "is it in the config", description
formatting, and SafeAutoCorrect checks. Now matched on cop name instead.

Verified end to end that a project inheriting only default.yml does not run the cop
and one that also inherits sorbet.yml does. Re-ran the legacy-vs-upstream comparison
over 12,495 zenpayroll files after the rename: 40 files flagged, zero divergences in
offenses or autocorrect output.
